requirements.txt
fastapi==0.115.12
Jinja2==3.1.6
uvicorn==0.34.2
annotated-types==0.7.0
anyio==4.9.0
certifi==2025.4.26
charset-normalizer==3.4.2
click==8.2.0
colorama==0.4.6
exception==0.1.0
fastapi==0.115.12
filelock==3.13.1
fsspec==2024.6.1
h11==0.16.0
huggingface-hub==0.31.2
idna==3.10
lxml==5.4.0
MarkupSafe==2.1.5
mpmath==1.3.0
networkx==3.3
numpy==2.2.5
opencv-python==4.11.0.86
packaging==25.0
pillow==11.0.0
pydantic==2.11.4
pydantic_core==2.33.2
python-docx==1.1.2
python-multipart==0.0.20
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.3
safetensors==0.5.3
six==1.17.0
sniffio==1.3.1
starlette==0.46.2
sympy==1.13.3
tokenizers==0.21.1
torch==2.7.0+cu126
torchaudio==2.7.0+cu126
torchvision==0.22.0+cu126
tqdm==4.67.1
transformers==4.51.3
typing-inspection==0.4.0
typing_extensions==4.13.2
urllib3==2.4.0
httpx
python-multipart
